refactor(terminal): extract TerminalSettleWaiterRegistry
- Move the readiness/attach-settled/fully-settled waiter maps and their
  waitFor/notify/remove/reject-all logic into a new sibling controller
- TerminalInstanceService keeps thin public delegates (waitForInstance,
  waitForAttachSettled, waitForFullySettled, waitForAllFullySettled,
  notifyRestoreSettledWaiters) plus the internal notify call sites
- Update tests that reached the moved private members directly
